server/requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-socketio==5.5.0
eventlet==0.35.2
google-generativeai==0.8.5
python-dotenv==1.0.1
pymysql==1.1.0
